+++ mayaLib/pipelineLib/utility/libManager.py
@@ -14,6 +14,7 @@ class InstallLibrary(QObject):
     def __init__(self, devMode=True, parent=None):
         super(InstallLibrary, self).__init__(parent)
 
+        self.devMode = devMode
         self.libUrl = 'https://github.com/Aiacos/DevPyLib/archive/master.zip'
         self.homeUser = os.getenv("HOME")
         self.mayaScriptPath = self.homeUser + '/Library/Preferences/Autodesk/maya/scripts/'
@@ -71,8 +72,10 @@ else:
             print 'ERROR: Directory not exist!'
 
     def install(self):
-        self.download()
+        if not self.devMode:
+            self.download()
         self.installInMayaUserSetup()
+        reload(__import__(self.libName))
 
     def uninstall(self):
         userSetup_path = self.mayaScriptPath
@@ -86,7 +89,6 @@ else:
 
                 f = open(filePath, 'w')
                 script = script.replace(self.installCommand, '')
-                print script
                 f.write(script)
                 f.close()
 
@@ -101,7 +103,7 @@ else:
         progress_size = int(count * block_size)
         speed = int(progress_size / (1024 * duration))
         percent = int(count * block_size * 100 / total_size)
-        sys.stdout.write("\r...%d%%, %d MB, %d KB/s, %d seconds passed" %
+        sys.stdout.write("\r...%d%%, %d MB, %d KB/s, %d seconds passed\n" %
                          (percent, progress_size / (1024 * 1024), speed, duration))
         sys.stdout.flush()
 
@@ -127,8 +129,9 @@ else:
     def delete(self):
         cd_cmd = 'cd ' + self.mayaScriptPath + ' && '
         # remove
-        rm_cmd = cd_cmd + 'rm -R DevPyLib-master'
-        os.system(rm_cmd)
+        if os.path.isdir(self.mayaScriptPath + 'DevPyLib-master'):
+            rm_cmd = cd_cmd + 'rm -R DevPyLib-master'
+            os.system(rm_cmd)
 
 class InstallWindow(QWidget):
     def __init__(self, parent=None):
@@ -158,7 +161,6 @@ class InstallWindow(QWidget):
         self.installDirLineEdit.setText(QFileDialog.getExistingDirectory(self, "Select Directory"))
 
     def installLib(self):
-        print 'install clicked'
         self.libManager.install()
